fix some issues in delte review

+++ src/controllers/reviews/deletereview.ts
@@ -5,13 +5,20 @@ import Review from '../../database/reviews';
 
 const deleteReview = async (req: TokenRequest, res: Response, next: NextFunction):Promise<void> => {
   try {
-    console.log(req.body, req.user);
     const { id } = req.user;
-    const { productID } = req.body;
-    await Review.deleteOne({
+    const { productID } = req.params;
+    const reviewExist = await Review.findOne({
       userID: id,
       productID,
-    });
+    })
+    if (!reviewExist) {
+      next( new CustomError(400, "Review doesn't exist!"))
+    } else {
+      await Review.deleteOne({
+        userID: id,
+        productID,
+      });
+    }
     res.status(200).json({ Massage: 'Review deleted successfully' });
   } catch (error) {
     next(new CustomError(500, error.massage));
